Add files via upload
Cisco verifies the earner of this badge successfully completed the Switching, Routing, and Wireless Essentials course and achieved this student level credential. The earner has a foundation in:
•	Switching operations
•	Wired and wireless LAN configuration using security best practices
•	Redundancy protocols
•	Problem-solving skills
The earner participated in up to 45 practice activities and accumulated up to 21 hours of hands-on labs using Cisco hardware and/or the Cisco Packet Tracer tool.