Add files via upload
The script Reset-HybridADJoinDevice unjoins a Windows machine from Azure AD, removes leftover hybrid join certificates, and then re-initiates the join process using the built-in scheduled task "Automatic-Device-Join". It validates the success of the operation by checking for certificate creation and join status. Optionally, it can be run against a list of computer names.